base/modules/tmp/sb_utils/root/requirements.txt
bson==0.5.10
canonicaljson==1.4.0
cbor2==5.2.0
cryptography==3.1.1
jws==0.1.3
kombu==5.0.2
msgpack==1.0.0
py3-bencode==0.0.3
py-ubjson==0.16.1
pycrypto==2.6.1
pyrabbit==1.1.0
python-etcd==0.4.5
pyyaml==5.3.1
sexpdata==0.0.3
toml==0.10.1
xmltodict==0.12.0